pyproject.toml
[tool.poetry]
name = "plasticome"
version = "0.1.0"
description = ""
authors = ["Blueevee <lm.elyn@hotmail.com>"]
readme = "README.md"
packages = [{include = "plasticome"}]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.81"
taskipy = "^1.12.0"
flask-pydantic-spec = "^0.4.5"
python-dotenv = "^1.0.0"
flask = "^2.3.3"
celery = "^5.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
blue = "^0.9.1"
isort = "^5.12.0"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.1.18"
mkdocstrings = "^0.22.0"
mkdocstrings-python = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=plasticome -vv"
post_test = "coverage html"

